+++ src/elements/CardNumberElement.tsx
@@ -1,4 +1,4 @@
-import React, { FC } from 'react';
+import React, { FC, useRef, ForwardedRef } from 'react';
 import type {
   CardNumberElement as ICardNumberElement,
   CreateCardNumberElementOptions,
@@ -11,7 +11,7 @@ import type { BasisTheoryReact } from '../core';
 import { useElement } from './useElement';
 import { useListener } from './useListener';
 
-export interface CardNumberElementProps {
+interface CardNumberElementProps {
   id: string;
   bt?: BasisTheoryReact;
   style?: ElementStyle;
@@ -25,9 +25,10 @@ export interface CardNumberElementProps {
   onBlur?: ElementEventListener<CardNumberElementEvents, 'blur'>;
   onReady?: ElementEventListener<CardNumberElementEvents, 'ready'>;
   onKeyDown?: ElementEventListener<CardNumberElementEvents, 'keydown'>;
+  elementRef?: ForwardedRef<ICardNumberElement>;
 }
 
-export const CardNumberElement: FC<CardNumberElementProps> = ({
+const CardNumberElementC: FC<CardNumberElementProps> = ({
   id,
   bt,
   style,
@@ -41,13 +42,16 @@ export const CardNumberElement: FC<CardNumberElementProps> = ({
   onFocus,
   onBlur,
   onKeyDown,
+  elementRef,
 }) => {
+  const wrapperRef = useRef<HTMLDivElement>(null);
   const element = useElement<
     ICardNumberElement,
     CreateCardNumberElementOptions
   >(
     id,
     'cardNumber',
+    wrapperRef,
     {
       targetId: id,
       style,
@@ -57,7 +61,8 @@ export const CardNumberElement: FC<CardNumberElementProps> = ({
       placeholder,
       iconPosition,
     },
-    bt
+    bt,
+    elementRef
   );
 
   useListener('ready', element, onReady);
@@ -66,5 +71,15 @@ export const CardNumberElement: FC<CardNumberElementProps> = ({
   useListener('blur', element, onBlur);
   useListener('keydown', element, onKeyDown);
 
-  return <div id={id} />;
+  return <div id={id} ref={wrapperRef} />;
 };
+
+export const CardNumberElement = React.forwardRef<
+  ICardNumberElement,
+  CardNumberElementProps
+  // eslint-disable-next-line get-off-my-lawn/prefer-arrow-functions
+>(function CardNumberElement(props, ref) {
+  return <CardNumberElementC {...props} elementRef={ref} />;
+});
+
+export type { CardNumberElementProps };