+++ test/elements/CardElement.test.tsx
@@ -1,5 +1,8 @@
 import * as React from 'react';
-import type { ElementStyle } from '@basis-theory/basis-theory-js/types/elements';
+import type {
+  CardElement as ICardElement,
+  ElementStyle,
+} from '@basis-theory/basis-theory-js/types/elements';
 import { render } from '@testing-library/react';
 import { Chance } from 'chance';
 import { CardElement } from '../../src';
@@ -11,7 +14,10 @@ jest.mock('../../src/elements/useListener');
 
 describe('CardElement', () => {
   const chance = new Chance();
+  const refArray = [React.createRef<ICardElement>(), undefined];
 
+  let id: string;
+  let wrapperDiv: HTMLDivElement;
   let style: ElementStyle;
   let disabled: boolean;
   let autoComplete: string;
@@ -21,8 +27,12 @@ describe('CardElement', () => {
   let onBlur: jest.Mock;
   let onKeyDown: jest.Mock;
   let element: unknown;
+  let ref: any;
 
   beforeEach(() => {
+    id = 'my-card';
+    wrapperDiv = document.createElement('div');
+    wrapperDiv.setAttribute('id', id);
     style = {
       [chance.string()]: chance.string(),
     };
@@ -36,6 +46,7 @@ describe('CardElement', () => {
     element = {
       [chance.string()]: chance.string(),
     };
+    ref = chance.pickone(refArray);
 
     jest.mocked(useElement).mockReturnValue(element as any);
   });
@@ -45,26 +56,30 @@ describe('CardElement', () => {
       <CardElement
         autoComplete={autoComplete}
         disabled={disabled}
-        id="my-card"
+        id={id}
         onBlur={onBlur}
         onChange={onChange}
         onFocus={onFocus}
         onKeyDown={onKeyDown}
         onReady={onReady}
+        ref={ref}
         style={style}
       />
     );
 
     expect(container).toMatchSnapshot();
     expect(useElement).toHaveBeenCalledWith(
-      'my-card',
+      id,
       'card',
+      { current: wrapperDiv },
       {
         style,
         disabled,
         autoComplete,
       },
-      undefined
+      undefined,
+      // eslint-disable-next-line unicorn/no-null
+      typeof ref === 'undefined' ? null : ref // undefined ref gets forwarded as null
     );
     expect(useListener).toHaveBeenCalledWith('ready', element, onReady);
     expect(useListener).toHaveBeenCalledWith('change', element, onChange);