+++ test/elements/CardExpirationDateElement.test.tsx
@@ -1,5 +1,8 @@
 import * as React from 'react';
-import type { ElementStyle } from '@basis-theory/basis-theory-js/types/elements';
+import type {
+  CardExpirationDateElement as ICardExpirationDateElement,
+  ElementStyle,
+} from '@basis-theory/basis-theory-js/types/elements';
 import { render } from '@testing-library/react';
 import { Chance } from 'chance';
 import { CardExpirationDateElement } from '../../src';
@@ -11,7 +14,10 @@ jest.mock('../../src/elements/useListener');
 
 describe('CardExpirationDateElement', () => {
   const chance = new Chance();
+  const refArray = [React.createRef<ICardExpirationDateElement>(), undefined];
 
+  let id: string;
+  let wrapperDiv: HTMLDivElement;
   let style: ElementStyle;
   let disabled: boolean;
   let autoComplete: string;
@@ -23,8 +29,12 @@ describe('CardExpirationDateElement', () => {
   let onBlur: jest.Mock;
   let onKeyDown: jest.Mock;
   let element: unknown;
+  let ref: any;
 
   beforeEach(() => {
+    id = 'my-card-expiration-date';
+    wrapperDiv = document.createElement('div');
+    wrapperDiv.setAttribute('id', id);
     style = {
       [chance.string()]: chance.string(),
     };
@@ -41,6 +51,7 @@ describe('CardExpirationDateElement', () => {
     element = {
       [chance.string()]: chance.string(),
     };
+    ref = chance.pickone(refArray);
 
     jest.mocked(useElement).mockReturnValue(element as any);
   });
@@ -51,30 +62,34 @@ describe('CardExpirationDateElement', () => {
         aria-label={ariaLabel}
         autoComplete={autoComplete}
         disabled={disabled}
-        id="my-card-expiration-date"
+        id={id}
         onBlur={onBlur}
         onChange={onChange}
         onFocus={onFocus}
         onKeyDown={onKeyDown}
         onReady={onReady}
         placeholder={placeholder}
+        ref={ref}
         style={style}
       />
     );
 
     expect(container).toMatchSnapshot();
     expect(useElement).toHaveBeenCalledWith(
-      'my-card-expiration-date',
+      id,
       'cardExpirationDate',
+      { current: wrapperDiv },
       {
-        targetId: 'my-card-expiration-date',
+        targetId: id,
         style,
         disabled,
         autoComplete,
         'aria-label': ariaLabel,
         placeholder,
       },
-      undefined
+      undefined,
+      // eslint-disable-next-line unicorn/no-null
+      typeof ref === 'undefined' ? null : ref // undefined ref gets forwarded as null
     );
     expect(useListener).toHaveBeenCalledWith('ready', element, onReady);
     expect(useListener).toHaveBeenCalledWith('change', element, onChange);