+++ test/elements/CardVerificationCodeElement.test.tsx
@@ -3,6 +3,7 @@ import { CARD_BRANDS } from '@basis-theory/basis-theory-js/elements/constants';
 import type {
   Brand,
   ElementStyle,
+  CardVerificationCodeElement as ICardVerificationCodeElement,
 } from '@basis-theory/basis-theory-js/types/elements';
 import { render } from '@testing-library/react';
 import { Chance } from 'chance';
@@ -15,7 +16,10 @@ jest.mock('../../src/elements/useListener');
 
 describe('CardVerificationCodeElement', () => {
   const chance = new Chance();
+  const refArray = [React.createRef<ICardVerificationCodeElement>(), undefined];
 
+  let id: string;
+  let wrapperDiv: HTMLDivElement;
   let style: ElementStyle;
   let disabled: boolean;
   let autoComplete: string;
@@ -28,8 +32,12 @@ describe('CardVerificationCodeElement', () => {
   let onBlur: jest.Mock;
   let onKeyDown: jest.Mock;
   let element: unknown;
+  let ref: any;
 
   beforeEach(() => {
+    id = 'my-card-verification-code';
+    wrapperDiv = document.createElement('div');
+    wrapperDiv.setAttribute('id', id);
     style = {
       [chance.string()]: chance.string(),
     };
@@ -47,6 +55,7 @@ describe('CardVerificationCodeElement', () => {
     element = {
       [chance.string()]: chance.string(),
     };
+    ref = chance.pickone(refArray);
 
     jest.mocked(useElement).mockReturnValue(element as any);
   });
@@ -58,23 +67,25 @@ describe('CardVerificationCodeElement', () => {
         autoComplete={autoComplete}
         cardBrand={cardBrand}
         disabled={disabled}
-        id="my-card-verification-code"
+        id={id}
         onBlur={onBlur}
         onChange={onChange}
         onFocus={onFocus}
         onKeyDown={onKeyDown}
         onReady={onReady}
         placeholder={placeholder}
+        ref={ref}
         style={style}
       />
     );
 
     expect(container).toMatchSnapshot();
     expect(useElement).toHaveBeenCalledWith(
-      'my-card-verification-code',
+      id,
       'cardVerificationCode',
+      { current: wrapperDiv },
       {
-        targetId: 'my-card-verification-code',
+        targetId: id,
         style,
         disabled,
         autoComplete,
@@ -82,7 +93,9 @@ describe('CardVerificationCodeElement', () => {
         placeholder,
         cardBrand,
       },
-      undefined
+      undefined,
+      // eslint-disable-next-line unicorn/no-null
+      typeof ref === 'undefined' ? null : ref // undefined ref gets forwarded as null
     );
     expect(useListener).toHaveBeenCalledWith('ready', element, onReady);
     expect(useListener).toHaveBeenCalledWith('change', element, onChange);