+++ test/elements/TextElement.test.tsx
@@ -1,5 +1,8 @@
 import * as React from 'react';
-import type { ElementStyle } from '@basis-theory/basis-theory-js/types/elements';
+import type {
+  TextElement as ITextElement,
+  ElementStyle,
+} from '@basis-theory/basis-theory-js/types/elements';
 import { render } from '@testing-library/react';
 import { Chance } from 'chance';
 import { TextElement } from '../../src';
@@ -11,7 +14,10 @@ jest.mock('../../src/elements/useListener');
 
 describe('TextElement', () => {
   const chance = new Chance();
+  const refArray = [React.createRef<ITextElement>(), undefined];
 
+  let id: string;
+  let wrapperDiv: HTMLDivElement;
   let style: ElementStyle;
   let disabled: boolean;
   let autoComplete: string;
@@ -26,8 +32,12 @@ describe('TextElement', () => {
   let onBlur: jest.Mock;
   let onKeyDown: jest.Mock;
   let element: unknown;
+  let ref: any;
 
   beforeEach(() => {
+    id = 'my-input';
+    wrapperDiv = document.createElement('div');
+    wrapperDiv.setAttribute('id', id);
     style = {
       [chance.string()]: chance.string(),
     };
@@ -53,6 +63,7 @@ describe('TextElement', () => {
     element = {
       [chance.string()]: chance.string(),
     };
+    ref = chance.pickone(refArray);
 
     jest.mocked(useElement).mockReturnValue(element as any);
   });
@@ -63,7 +74,7 @@ describe('TextElement', () => {
         aria-label={ariaLabel}
         autoComplete={autoComplete}
         disabled={disabled}
-        id="my-input"
+        id={id}
         mask={mask as never} // the need of this cast acts a props typings test :-)
         onBlur={onBlur}
         onChange={onChange}
@@ -72,6 +83,7 @@ describe('TextElement', () => {
         onReady={onReady}
         password={password}
         placeholder={placeholder}
+        ref={ref}
         style={style}
         transform={transform}
       />
@@ -79,10 +91,11 @@ describe('TextElement', () => {
 
     expect(container).toMatchSnapshot();
     expect(useElement).toHaveBeenCalledWith(
-      'my-input',
+      id,
       'text',
+      { current: wrapperDiv },
       {
-        targetId: 'my-input',
+        targetId: id,
         style,
         disabled,
         autoComplete,
@@ -92,7 +105,9 @@ describe('TextElement', () => {
         placeholder,
         transform,
       },
-      undefined
+      undefined,
+      // eslint-disable-next-line unicorn/no-null
+      typeof ref === 'undefined' ? null : ref // undefined ref gets forwarded as null
     );
     expect(useListener).toHaveBeenCalledWith('ready', element, onReady);
     expect(useListener).toHaveBeenCalledWith('change', element, onChange);