+++ test/elements/useElement.test.ts
@@ -34,9 +34,26 @@ describe('useElement', () => {
   });
 
   test("shouldn't do anything while bt instance is not available", () => {
+    const mockRef = { current: document.createElement('div') };
+
     jest.mocked(useBasisTheoryValue).mockReturnValue(undefined);
+
     const { result } = renderHook(() =>
-      useElement(chance.string(), chance.pickone(['card', 'text']), {})
+      useElement(chance.string(), chance.pickone(['card', 'text']), mockRef, {})
+    );
+
+    expect(bt.createElement).toHaveBeenCalledTimes(0);
+    expect(bt.indexElement).toHaveBeenCalledTimes(0);
+    expect(bt.disposeElement).toHaveBeenCalledTimes(0);
+    expect(result.current).toBeUndefined();
+  });
+
+  test("shouldn't do anything while wrapper ref is not in the DOM", () => {
+    // eslint-disable-next-line unicorn/no-null
+    const mockRef = { current: null };
+
+    const { result } = renderHook(() =>
+      useElement(chance.string(), chance.pickone(['card', 'text']), mockRef, {})
     );
 
     expect(bt.createElement).toHaveBeenCalledTimes(0);
@@ -48,7 +65,12 @@ describe('useElement', () => {
   test('should create, mount and index element', () => {
     const id = chance.string();
     const type = chance.pickone<ElementType>(['card', 'text']);
-    const { result } = renderHook(() => useElement(id, type, {}));
+    const mockRef = { current: document.createElement('div') };
+    // eslint-disable-next-line unicorn/no-null
+    const testRef = { current: null };
+    const { result } = renderHook(() =>
+      useElement(id, type, mockRef, {}, undefined, testRef)
+    );
 
     expect(bt.createElement).toHaveBeenCalledTimes(1);
     expect(bt.createElement).toHaveBeenCalledWith(type, {});
@@ -58,15 +80,50 @@ describe('useElement', () => {
     expect(bt.indexElement).toHaveBeenCalledWith(id, result.current);
   });
 
+  test('should forward object ref to element on creation', () => {
+    const id = chance.string();
+    const type = chance.pickone<ElementType>(['card', 'text']);
+    const wrapperRef = { current: document.createElement('div') };
+
+    // eslint-disable-next-line unicorn/no-null
+    const objectRef = { current: null };
+    const { result } = renderHook(() =>
+      useElement(id, type, wrapperRef, {}, undefined, objectRef)
+    );
+
+    expect(result.current).toBeDefined();
+    expect(objectRef.current).toBe(result.current);
+  });
+
+  test('should forward function ref to element on creation', () => {
+    const id = chance.string();
+    const type = chance.pickone<ElementType>(['card', 'text']);
+    const wrapperRef = { current: document.createElement('div') };
+
+    // eslint-disable-next-line unicorn/no-null
+    let createdElement = null;
+    const functionRef = (element: any): void => {
+      createdElement = element;
+    };
+
+    const { result } = renderHook(() =>
+      useElement(id, type, wrapperRef, {}, undefined, functionRef)
+    );
+
+    expect(result.current).toBeDefined();
+    expect(createdElement).toBe(result.current);
+  });
+
   test('should throw mount error in lifecycle', async () => {
     const id = chance.string();
     const type = chance.pickone<ElementType>(['card', 'text']);
+    const mockRef = { current: document.createElement('div') };
     const errorMessage = chance.string();
 
     mount.mockRejectedValue(errorMessage);
 
     const { result, waitForNextUpdate } = renderHook(() =>
-      useElement(id, type, {})
+      useElement(id, type, mockRef, {})
     );
 
     await waitForNextUpdate();
@@ -83,9 +140,10 @@ describe('useElement', () => {
 
     const id = chance.string();
     const type = chance.pickone<ElementType>(['card', 'text']);
+    const mockRef = { current: document.createElement('div') };
 
     const { rerender } = renderHook(
-      ({ options }) => useElement(id, type, options),
+      ({ options }) => useElement(id, type, mockRef, options),
       {
         initialProps: {
           options: {
@@ -154,12 +212,13 @@ describe('useElement', () => {
   test('should throw update error in lifecycle', async () => {
     const id = chance.string();
     const type = chance.pickone<ElementType>(['card', 'text']);
+    const mockRef = { current: document.createElement('div') };
     const errorMessage = chance.string();
 
     update.mockRejectedValue(errorMessage);
 
     const { result, rerender, waitForNextUpdate } = renderHook(
-      ({ options }) => useElement(id, type, options),
+      ({ options }) => useElement(id, type, mockRef, options),
       {
         initialProps: {
           options: {},